==> 01_Data_Sampling_Tim.ipynb <==
import pandas as pd 
import numpy as np

# local variables
path = 'data/raw_data/'
sample_data_path = 'data/sample_data/'

# user data
user = pd.read_csv(path + 'users.tsv',sep='\t')
# application data
apps = pd.read_csv(path + 'apps.tsv',sep='\t')
# jobs data
jobs = pd.read_csv(path + 'jobs.tsv',sep='\t',error_bad_lines=False)
# user history data 
user_history= pd.read_csv(path + 'user_history.tsv',sep='\t')

sample_size = 1000
sample_users = user.head(sample_size)
sample_users_id = []
for row in sample_users.itertuples():
    sample_users_id.append(row.UserID)


apps_sample = apps.loc[apps['UserID'].isin(sample_users_id)]
apps_sample

sample_jobs_id = []
for row in apps_sample.itertuples():
    sample_jobs_id.append(row.JobID)

#jobs_sample = jobs.loc[apps['JobID'].isin(sample_jobs_id)] #this is wrong
jobs_sample = jobs[jobs['JobID'].isin(sample_jobs_id)]

# there are jobs that user apply in app.tsv does not belong to jobs
jobs_job_id=[]
for jobID in jobs_sample['JobID']:
    jobs_job_id.append(jobID)

print("total data before remove unmatched jobs ",len(apps_sample))
apps_sample_clean = apps_sample.loc[apps_sample['JobID'].isin(jobs_job_id)]
print("total data after remove unmatched jobs ", len(apps_sample_clean))

# save to file
apps_sample.to_csv(sample_data_path + 'apps_sample.csv')

jobs_sample.to_csv(sample_data_path + 'jobs_sample.csv')

sample_users.to_csv(sample_data_path + 'users_sample.csv')

user_history_sample = user_history.loc[user_history['UserID'].isin(sample_users_id)]
user_history_sample

user_history_sample.to_csv(sample_data_path + 'user_history_sample.csv')
