==> 02_Data Clearing_Kun.ipynb <==
import pandas as pd 
import numpy as np

# local variables
sample_data_path = 'data/sample_data/'

# user data
users = pd.read_csv(sample_data_path + 'users_sample.csv')
# application data
apps = pd.read_csv(sample_data_path + 'apps_sample.csv')
# jobs data
jobs = pd.read_csv(sample_data_path + 'jobs_sample.csv')
# user history data 
user_history = pd.read_csv(sample_data_path + 'user_history_sample.csv')

# Find 10 most popular skills per position

JD = jobs[['JobID','Title','Description','Requirements','State','City']]
JD

## Data Clearing

### Removing urls

### reference: https://www.kaggle.com/shahules/basic-eda-cleaning-and-glove
JD_Description = JD['Description']
example = JD_Description[0]
example

import re

def remove_URL(text):
    url = re.compile(r'(https?:\/\/)?[-a-zA-Z0-9@:%_\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}([-a-zA-Z0-9()@:%_\+.~#?&=]*)\/([-a-zA-Z0-9()@:%_\+.~#?&=]*)')
    return url.sub(r'',text)
example = remove_URL(example)
example


### Removing HTML tags

import re
import html

def remove_html(text):
    html=re.compile(r'<.*?>')
    return html.sub(r'',text)
example = remove_html(example)
example

### Removing Spaces

def remove_spaces(text):
    html=re.compile(r'&([-a-zA-Z0-9()@:%_\+.~#?&=]*);')
    return html.sub(r'',text)
example = remove_spaces(example)
example

# Not doing spellchecker at this time since it is way too slow

# from spellchecker import SpellChecker

# spell = SpellChecker()
# def correct_spellings(text):
#     corrected_text = []
#     misspelled_words = spell.unknown(text.split())
#     for word in text.split():
#         if word in misspelled_words:
#             corrected_text.append(spell.correction(word))
#         else:
#             corrected_text.append(word)
#     return " ".join(corrected_text)

# text = "I alwasy hsa smoe typo"
# correct_spellings(text)


### Combined all together

combined_example = JD_Description[10]
combined_example

def data_clearing_combined(text):
    
    regex = r'(https?:\/\/)?[-a-zA-Z0-9@:%_\+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}([-a-zA-Z0-9()@:%_\+.~#?&=]*)\/([-a-zA-Z0-9()@:%_\+.~#?&=]*)|<.*?>|&([-a-zA-Z0-9()@:%_\+.~#?&=]*);|\\+r|\\+n'
    result = re.compile(regex)
    return result.sub(r'',text)
                                  
combined_example = data_clearing_combined(combined_example)
combined_example


# Clear JD description
JD['Description'] = JD['Description'].apply(lambda x : data_clearing_combined(str(x)))
JD['Description']

JD['Requirements'] = JD['Requirements'].apply(lambda x : data_clearing_combined(str(x)))
JD['Requirements']

JD

JD.to_csv(sample_data_path + 'jobs_sample_cleared.csv')
