==> 05_user_rating_profile_Taylor.ipynb <==
### Read From sampling data

import pandas as pd 
import numpy as np

# local variables
sample_data_path = 'sample_data/'

# user data
users = pd.read_csv(sample_data_path + 'users_sample.csv')
# application data
apps = pd.read_csv(sample_data_path + 'apps_sample.csv')
# jobs data
jobs = pd.read_csv(sample_data_path + 'jobs_sample.csv')
# user history data 
user_history = pd.read_csv(sample_data_path + 'user_history_sample.csv')

### Generating the user-rating-profile based on application history

#if users applied this job that means very interested, we assign score 10 to this job
user_profile = pd.get_dummies(apps.JobID).groupby(apps.UserID).apply(max)
user_profile.replace(to_replace =1, value =10)

### Save the sampled matrix data

#Final - get the user_profile_ratings
user_profile.to_csv(sample_data_path + 'user_profile.csv')

### Rearrange the data to UserID/JobID/rating

apps

users_rating = apps.loc[:,['UserID', 'JobID']] 

users_rating['rating'] = 5

users_rating.to_csv(sample_data_path + 'users_rating.csv')

users_rating

### Add the popular job info to the user-job-id
#### Based on pupularity.csv, assign a small score - 1 to users, which it helps solve the cold star problem

pop = pd.read_csv(sample_data_path + 'popular_jobs.csv')

pop

# revise the name
pop.rename(columns={' JobIds':'JobIds'}, inplace=True)
pop.dropna()

# remove the ones that is not a string
pop['rating'] = pop['JobIds'].apply(lambda x: type(x) is str)
pop = pop.loc[pop['rating'], :]

pop['JobIds'] = pop['JobIds'].apply(lambda x: x.split())


def getTOP10(x):
    if len(x)>10:
        x = x[:10]
    return x
pop['JobIds'] = pop['JobIds'].apply(lambda x: getTOP10(x))
pop['rating'] = pop['rating'].replace(to_replace =True, 
                 value =1) 

pop['JobIds']

lst_col = 'JobIds'

r = pd.DataFrame({
      col:np.repeat(pop[col].values, pop[lst_col].str.len())
      for col in pop.columns.drop(lst_col)}
    ).assign(**{lst_col:np.concatenate(pop[lst_col].values)})[pop.columns]

r

### Get the user_profile_ratings based on popularity 

#get the user_profile_ratings
r.to_csv(sample_data_path + 'pop_rating.csv')

r

### Combine the two user-app-rating and user-pop-rating togeter to the final user rating

#revise the name
r.rename(columns={'JobIds':'JobID'}, inplace=True)
r.rename(columns={'UserId':'UserID'}, inplace=True)

frames = [users_rating,r]
result = pd.concat(frames)

#Final - get the user_profile_ratings
result.to_csv(sample_data_path + 'new_user_rating.csv')

### Future work (Master Project)

### Second step

##### Build a job-job similarity based on titles
#### Assign the similar jobs that users applied to 1 score


### Third step

#### find the similar jobs to user_history
#### assign the similar jobs that users had to 1 score
