==> 06_jobRecommend_Taylor.ipynb <==
**Environment set up requiremnt** 

Please Use colab.google to run this since some of the required enviroment set up is needed

if related environment has already being setted up its okay to commend out all the lines from this section start with !

!pip install pyspark
!apt-get install openjdk-8-jdk-headless -qq > /dev/null

import os

os.environ["JAVA_HOME"] = "/usr/lib/jvm/java-8-openjdk-amd64" 

!update-alternatives --set java /usr/lib/jvm/java-8-openjdk-amd64/jre/bin/java 
!java -version

from __future__ import print_function

import sys

if sys.version >= '3':
    long = int

from pyspark.sql import SparkSession
from pyspark import SparkContext, SparkConf

from pyspark.ml.evaluation import RegressionEvaluator
from pyspark.ml.recommendation import ALS
from pyspark.sql import Row
import csv
import pyspark.sql.functions as f

### Using spark.ml to Generate the result
#### need to upload the required csv file from previous user-profile-ipynb 

sc = SparkContext("local", "RecommendationSystem")
spark = SparkSession(sc)

lines = spark.read.text("./new_user_rating.csv").rdd

### Get rid of the hearder

header = lines.first()
header = sc.parallelize([header]) 
lines = lines.subtract(header)

parts = lines.map(lambda row: row.value.split(","))
ratingsRDD = parts.map(lambda p: Row(userId=int(p[1]), jobId=int(p[2]),rating=float(p[3])))
ratings = spark.createDataFrame(ratingsRDD)

### Split to training and test to a ritio of 8:2
#### Using Explicit method

(training, test) = ratings.randomSplit([0.8, 0.2])
tpanda = training.toPandas()
testpand = test.toPandas()
# Build the recommendation model using ALS on the training data
# Set cold start strategy to 'drop' to ensure we don't get NaN evaluation metrics
# Change implict to true
als = ALS(maxIter=5, regParam=0.01, userCol="userId", itemCol="jobId", ratingCol="rating",
          coldStartStrategy="drop")
model = als.fit(training)

# Evaluate the model by computing the RMSE on the test data
predictions = model.transform(test)
evaluator = RegressionEvaluator(metricName="rmse", labelCol="rating",
                                predictionCol="prediction")
rmse = evaluator.evaluate(predictions)
print("Root-mean-square error = " + str(rmse))

# Generate top 10 job recommendations for each user
userRecs = model.recommendForAllUsers(5)

userRecs.show()

### Saving the result to csv file

userRecs.toPandas().to_csv('Top10ForUserwithExplict.csv')

### Using implicit method

(training, test) = ratings.randomSplit([0.8, 0.2])
tpanda = training.toPandas()
testpand = test.toPandas()
# Build the recommendation model using ALS on the training data
# Set cold start strategy to 'drop' to ensure we don't get NaN evaluation metrics
# Change implict to true
als = ALS(maxIter=5, regParam=0.01, implicitPrefs=True, userCol="userId", itemCol="jobId", ratingCol="rating",
          coldStartStrategy="drop")
model = als.fit(training)

# Evaluate the model by computing the RMSE on the test data
predictions = model.transform(test)
evaluator = RegressionEvaluator(metricName="rmse", labelCol="rating",
                                predictionCol="prediction")
rmse = evaluator.evaluate(predictions)
print("Root-mean-square error = " + str(rmse))

# Generate top 10 job recommendations for each user
userRecs = model.recommendForAllUsers(5)

userRecs.show()

### Saving the result to csv file

userRecs.toPandas().to_csv('Top10ForUserwithImplicit.csv')
